Task completed, just need to double check things and write a ReadMe file

+++ db/migrate/20221230173927_create_orders.rb
@@ -2,8 +2,8 @@ class CreateOrders < ActiveRecord::Migration[7.0]
   def change
     create_table :orders do |t|
       t.integer :user_id
-      t.integer :quantity
-      t.integer :price
+      t.float :quantity
+      t.float :price
       t.integer :otype
 
       t.timestamps
